# src/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a programming assignment for GWI Go Backend position",
        "title": "GWI Go Challenge API",
        "contact": {
            "name": "Dionysis Chasakis",
            "email": "chasakisd@gmail.com"
        },
        "version": "1.0"
    },
    "basePath": "/",
    "paths": {
        "/asset": {
            "get": {
                "description": "Returns all the application assets",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Asset"
                ],
                "summary": "Get assets",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Page",
                        "name": "page",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.AssetResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/asset/favorite": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Get user's favorite assets",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Asset"
                ],
                "summary": "Get favorite assets",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Page",
                        "name": "page",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.AssetResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": ""
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "response.AssetResponse": {
            "type": "object",
            "properties": {
                "audiences": {
                    "$ref": "#/definitions/response.AudiencePageResponse"
                },
                "charts": {
                    "$ref": "#/definitions/response.ChartPageResponse"
                },
                "insights": {
                    "$ref": "#/definitions/response.InsightPageResponse"
                }
            }
        },
...
        "response.AudiencePageResponse": {
            "type": "object",
            "properties": {
                "page": {
                    "type": "integer"
                },
                "results": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/response.AudienceResponse"
                    }
                }
            }
        },
        "response.AudienceResponse": {
            "type": "object",
            "properties": {
                "ageGroupMax": {
                    "type": "integer"
                },
                "ageGroupMin": {
                    "type": "integer"
                },
                "assetId": {
                    "type": "string"
                },
                "birthCountry": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "descriptionFormatted": {
                    "type": "string"
                },
                "gender": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "statType": {
                    "$ref": "#/definitions/response.AudienceStatTypeResponse"
                },
                "statTypeValue": {
                    "type": "number"
                }
            }
        },
...
        "response.AudienceStatTypeResponse": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                },
                "titleFormatted": {
                    "type": "string"
                }
            }
        },
...
        "response.ChartPageResponse": {
            "type": "object",
            "properties": {
                "page": {
                    "type": "integer"
                },
                "results": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/response.ChartResponse"
                    }
                }
            }
        },
...
        "response.ChartPointResponse": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                },
                "xValue": {
                    "type": "number"
                },
                "yValue": {
                    "type": "number"
                }
            }
        },
        "response.ChartResponse": {
            "type": "object",
            "properties": {
                "assetId": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "points": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/response.ChartPointResponse"
                    }
                },
                "xAxes": {
                    "type": "string"
                },
                "yAxes": {
                    "type": "string"
                }
            }
        },
        "response.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                }
            }
        },
...
        "response.InsightPageResponse": {
            "type": "object",
            "properties": {
                "page": {
                    "type": "integer"
                },
                "results": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/response.InsightResponse"
                    }
                }
            }
        },
        "response.InsightResponse": {
            "type": "object",
            "properties": {
                "assetId": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}